+++ core/src/subscription/features_s_bytes_sum.rs
@@ -89,7 +89,7 @@ pub struct TrackedFeatures {
     #[cfg(feature = "timing")]
     compute_ns: u64,
     cnt: u64,
-    s_bytes_sum: i64,
+    s_bytes_sum: f64,
 }
 
 impl TrackedFeatures {
@@ -104,7 +104,7 @@ impl TrackedFeatures {
         let ipv4 = eth.parse_to::<Ipv4>()?;
 
         if segment.dir {
-            self.s_bytes_sum += ipv4.total_length() as i64;
+            self.s_bytes_sum += ipv4.total_length() as f64;
         } else {
         }
         #[cfg(feature = "timing")]
@@ -119,7 +119,7 @@ impl TrackedFeatures {
     fn extract_features(&mut self) -> Vec<f64> {
         #[cfg(feature = "timing")]
         let start_ts = (unsafe { rte_rdtsc() } as f64 / *TSC_GHZ) as u64;
-        let features = vec![self.s_bytes_sum as f64];
+        let features = vec![self.s_bytes_sum];
         #[cfg(feature = "timing")]
         {
             let end_ts = (unsafe { rte_rdtsc() } as f64 / *TSC_GHZ) as u64;
@@ -137,7 +137,7 @@ impl Trackable for TrackedFeatures {
             #[cfg(feature = "timing")]
             compute_ns: 0,
             cnt: 0,
-            s_bytes_sum: 0,
+            s_bytes_sum: 0.0,
         }
     }
